# openapi.yaml
components:
  schemas:
...
    Body_create_evidence_api_v1_evidence__post:
      properties:
        description:
          default: ''
          title: Description
          type: string
        file:
          format: binary
          title: File
          type: string
        investigator_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Investigator Id
        risk_level:
          default: medium
          title: Risk Level
          type: string
        tags:
          default: ''
          title: Tags
          type: string
        title:
          default: ''
          title: Title
          type: string
        wallet_id:
          title: Wallet Id
          type: string
      required:
      - file
      - wallet_id
      title: Body_create_evidence_api_v1_evidence__post
      type: object
    Body_login_api_v1_auth_login_post:
      properties:
        client_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Client Id
        client_secret:
          anyOf:
          - type: string
          - type: 'null'
          title: Client Secret
        grant_type:
          anyOf:
          - pattern: password
            type: string
          - type: 'null'
          title: Grant Type
        password:
          title: Password
          type: string
        scope:
          default: ''
          title: Scope
          type: string
        username:
          title: Username
          type: string
      required:
      - username
      - password
      title: Body_login_api_v1_auth_login_post
      type: object
...
    EvidenceResponse:
      properties:
        anchor_status:
          anyOf:
          - type: string
          - type: 'null'
          default: pending
          title: Anchor Status
        case_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Case Id
        created_at:
          format: date-time
          title: Created At
          type: string
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        evidence_id:
          title: Evidence Id
          type: string
        file_path:
          anyOf:
          - type: string
          - type: 'null'
          title: File Path
        file_size:
          anyOf:
          - type: integer
          - type: 'null'
          title: File Size
        file_type:
          anyOf:
          - type: string
          - type: 'null'
          title: File Type
        hash:
          title: Hash
          type: string
        id:
          title: Id
          type: integer
        immutable:
          anyOf:
          - type: boolean
          - type: 'null'
          default: true
          title: Immutable
        investigator_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Investigator Id
        title:
          title: Title
          type: string
      required:
      - evidence_id
      - title
      - description
      - hash
      - file_path
      - file_size
      - file_type
      - anchor_status
      - immutable
      - id
      - investigator_id
      - case_id
      - created_at
      title: EvidenceResponse
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    Token:
      properties:
        access_token:
          title: Access Token
          type: string
        token_type:
          title: Token Type
          type: string
      required:
      - access_token
      - token_type
      title: Token
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
  securitySchemes:
    OAuth2PasswordBearer:
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login
      type: oauth2
info:
  description: Backend API for cybercrime investigation dashboard
  title: Cybercrime Investigation API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /api/v1/auth/login:
    post:
      description: Login and get access token
      operationId: login_api_v1_auth_login_post
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Body_login_api_v1_auth_login_post'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Login
      tags:
      - authentication
...
  /api/v1/evidence/:
    get:
      description: Get all evidence
      operationId: get_evidence_api_v1_evidence__get
      parameters:
      - in: query
        name: case_id
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Case Id
      - in: query
        name: investigator_id
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Investigator Id
      - in: query
        name: skip
        required: false
        schema:
          default: 0
          title: Skip
          type: integer
      - in: query
        name: limit
        required: false
        schema:
          default: 100
          title: Limit
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/EvidenceResponse'
                title: Response Get Evidence Api V1 Evidence  Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Evidence
      tags:
      - evidence
    post:
      description: 'Create a new evidence entry from an uploaded file.


        This is designed to work with the Investigator UI "Evidence Upload"

        section. It accepts ANY file type (PDF, images, CSV, Excel, docs, etc.)

        via multipart/form-data and stores the file plus metadata in SQL.

        Notifies superadmin when evidence is uploaded.'
      operationId: create_evidence_api_v1_evidence__post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_create_evidence_api_v1_evidence__post'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvidenceResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Evidence
      tags:
      - evidence
  /api/v1/evidence/{evidence_id}:
    get:
      description: Get a specific evidence item by ID
      operationId: get_evidence_item_api_v1_evidence__evidence_id__get
      parameters:
      - in: path
        name: evidence_id
        required: true
        schema:
          title: Evidence Id
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvidenceResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Evidence Item
      tags:
      - evidence
  /api/v1/evidence/{evidence_id}/download:
    get:
      description: Download an evidence file
      operationId: download_evidence_api_v1_evidence__evidence_id__download_get
      parameters:
      - in: path
        name: evidence_id
        required: true
        schema:
          title: Evidence Id
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Download Evidence
      tags:
      - evidence
  /api/v1/evidence/{evidence_id}/view:
    get:
      description: View an evidence file (for images, PDFs, etc.)
      operationId: view_evidence_api_v1_evidence__evidence_id__view_get
      parameters:
      - in: path
        name: evidence_id
        required: true
        schema:
          title: Evidence Id
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: View Evidence
      tags:
      - evidence
...